# swagger.json
{"openapi":"3.0.0","paths":{"/aaa":{"get":{"operationId":"AppController_aaa","parameters":[],"responses":{"200":{"description":""}}}},"/bbb":{"get":{"operationId":"AppController_bbb","parameters":[],"responses":{"200":{"description":""}}}},"/user/list":{"get":{"operationId":"UserController_list","summary":"获取全部用户","parameters":[{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageNo","required":true,"in":"query","schema":{"type":"number"}},{"name":"username","required":true,"in":"query","schema":{"type":"string"}},{"name":"nickName","required":true,"in":"query","schema":{"type":"string"}},{"name":"email","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/user/register":{"post":{"operationId":"UserController_create","summary":"用户注册","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserDto"}}}}},"tags":["用户"]}},"/user/register-captcha":{"get":{"operationId":"UserController_captcha","summary":"用户注册验证码","parameters":[{"name":"address","required":true,"in":"query","schema":{"type":"string"}},{"name":"","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["用户"]}},"/user/login":{"post":{"operationId":"UserController_login","summary":"用户登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}}},"tags":["用户"]}},"/user/refresh":{"get":{"operationId":"UserController_refresh","summary":"token刷新","parameters":[{"name":"refreshToken","required":true,"in":"query","schema":{"type":"string"}},{"name":"","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/user/info":{"get":{"operationId":"UserController_getInfo","summary":"获取用户信息","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDetailVo"}}}}},"tags":["用户"]}},"/user/updatePassword":{"post":{"operationId":"UserController_updatePassword","summary":"修改密码","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["用户"]}},"/user/updatePassword/sendCaptcha":{"get":{"operationId":"UserController_updatePasswordSendCaptcha","summary":"发送修改密码的验证码","parameters":[{"name":"address","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/user/update":{"post":{"operationId":"UserController_update","summary":"更改用户信息","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["用户"]}},"/user/update/sendCaptcha":{"get":{"operationId":"UserController_updateCaptcha","summary":"发送修改用户信息的验证码","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"bearer":[]}]}},"/user/freeze":{"get":{"operationId":"UserController_freeze","summary":"冻结用户","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/user/upload/headPic":{"post":{"operationId":"UserController_uploadFile","summary":"头像上传","parameters":[],"responses":{"201":{"description":""}},"tags":["用户"]}},"/meeting/list":{"get":{"operationId":"MeetingController_list","summary":"会议室列表","parameters":[{"name":"pageNo","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"pageSize","required":false,"in":"query","schema":{"maximum":50,"default":10,"type":"number"}},{"name":"name","required":false,"in":"query","description":"名称","schema":{"type":"string"}},{"name":"capacity","required":false,"in":"query","description":"容量","schema":{"type":"number"}},{"name":"equipment","required":false,"in":"query","description":"设备","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["会议室"]}},"/meeting/create":{"post":{"operationId":"MeetingController_create","summary":"创建会议室","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMeetingRoomDto"}}}},"responses":{"201":{"description":""}},"tags":["会议室"]}},"/meeting/update":{"put":{"operationId":"MeetingController_update","summary":"更新会议室","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMeetingDto"}}}},"responses":{"200":{"description":""}},"tags":["会议室"]}},"/meeting/{id}":{"get":{"operationId":"MeetingController_find","summary":"会议室详情","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["会议室"]},"delete":{"operationId":"MeetingController_delete","summary":"删除会议室","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["会议室"]}}},"info":{"title":"管理后台","description":"管理后台接口文档","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"basic":{"type":"http","scheme":"basic"}},"schemas":{"RegisterUserDto":{"type":"object","properties":{"username":{"type":"string","description":"用户名"},"nickName":{"type":"string","description":"昵称"},"password":{"type":"string","description":"密码"},"email":{"type":"string","description":"邮箱"},"captcha":{"type":"string","description":"验证码"}},"required":["username","nickName","password","email","captcha"]},"LoginUserDto":{"type":"object","properties":{"username":{"type":"string","description":"用户名"},"password":{"type":"string","description":"密码"}},"required":["username","password"]},"UserDetailVo":{"type":"object","properties":{}},"UpdateUserPasswordDto":{"type":"object","properties":{"username":{"type":"string","description":"账号"},"password":{"type":"string","description":"密码"},"email":{"type":"string","description":"邮箱"},"captcha":{"type":"string","description":"验证码"}},"required":["username","password","email","captcha"]},"UpdateUserDto":{"type":"object","properties":{"headPic":{"type":"string","description":"头像","nullable":true},"nickName":{"type":"string","description":"昵称","nullable":true},"email":{"type":"string","description":"邮箱"},"captcha":{"type":"string","description":"验证码"}},"required":["headPic","nickName","email","captcha"]},"CreateMeetingRoomDto":{"type":"object","properties":{"name":{"type":"string","description":"会议室名称"},"capacity":{"type":"number","description":"容量","uniqueItems":true},"location":{"type":"string","description":"位置"},"equipment":{"type":"string","description":"设备"},"description":{"type":"string","description":"描述","maximum":100}},"required":["name","capacity","location","equipment","description"]},"UpdateMeetingDto":{"type":"object","properties":{"name":{"type":"string","description":"会议室名称"},"capacity":{"type":"number","description":"容量","uniqueItems":true},"location":{"type":"string","description":"位置"},"equipment":{"type":"string","description":"设备"},"description":{"type":"string","description":"描述","maximum":100},"id":{"type":"number","description":"id"}},"required":["id"]}}}}